+++ app/models/unidom/product/concerns/as_source_product_associating.rb
@@ -1,0 +1,15 @@
+module Unidom::Product::Concerns::AsSourceProductAssociating
+
+  extend ActiveSupport::Concern
+
+  included do |includer|
+
+    has_many :target_product_associatings, class_name: 'Unidom::Product::ProductAssociating', foreign_key: :source_id
+    has_many :target_products,             through:    :target_product_associatings,          source: :target
+
+  end
+
+  module ClassMethods
+  end
+
+end